--- packages/dapp-kit-next/packages/dapp-kit-core/test/integration/integration-test-utils.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-export type TestWalletInitializeResult = {
-	unregister: () => void;
-};
